Improve TSDF mapper (high res color, feature, faster mesh extraction) (#694)
- Move feature and color from block level to patch level in tsdf grid. Each block can now store up to block size resolution of color and features. Makes it computationally faster compared to setting block_size=1 to get per voxel color/feature.
- Simplify mesh extraction, extraction is 10x faster from tsdf. Add projective texturing from color images when available.
- Fix bug in lidar tsdf example, fisheye projection is now done correctly to get points from unregistered pointcloud
- Minor: optimization kernels take int32 instead of int16

+++ curobo/_src/curobolib/backends/pybind/line_search_kernel_launch.cu
@@ -96,8 +96,8 @@
      <<< blocksPerGrid, threadsPerBlock, 0, stream >>> (
        best_cost.data_ptr<float>(),
        best_action.data_ptr<float>(),
-       best_iteration.data_ptr<int16_t>(),
-       current_iteration.data_ptr<int16_t>(),
+       best_iteration.data_ptr<int32_t>(),
+       current_iteration.data_ptr<int32_t>(),
        converged_global.data_ptr<uint8_t>(),
        convergence_iteration,
        cost_delta_threshold,
